requirements.txt
paddleocr
paddlepaddle
numpy
opencv-python
wheel
setuptools
PyMuPDF
gradio
torch 
torchvision 
torchaudio
realesrgan
Pillow